+++ src/gwt/src/org/rstudio/studio/client/common/mathjax/MathJax.java
@@ -49,7 +49,6 @@ import com.google.gwt.event.dom.client.BlurHandler;
 import com.google.gwt.event.logical.shared.AttachEvent;
 import com.google.gwt.event.shared.HandlerRegistration;
 import com.google.gwt.user.client.Command;
-import com.google.gwt.user.client.Window;
 import com.google.gwt.user.client.ui.FlowPanel;
 import com.google.inject.Inject;
 
@@ -67,20 +66,9 @@ public class MathJax
       docDisplay_ = docDisplay;
       popup_ = new MathJaxPopupPanel(this);
       bgRenderer_ = new MathJaxBackgroundRenderer(this, popup_, docDisplay_);
+      renderQueue_ = new MathJaxRenderQueue(this);
       handlers_ = new ArrayList<HandlerRegistration>();
       cowToPlwMap_ = new SafeMap<ChunkOutputWidget, PinnedLineWidget>();
-      onPopupTypeset_ = new MathJaxTypesetCallback()
-      {
-         @Override
-         public void onMathJaxTypesetComplete(boolean error)
-         {
-            int offsetWidth  = popup_.getOffsetWidth();
-            int offsetHeight = popup_.getOffsetHeight();
-            ScreenCoordinates coordinates = computePopupPosition(offsetWidth, offsetHeight);
-            popup_.setPopupPosition(coordinates.getPageX(), coordinates.getPageY());
-            popup_.show();
-         }
-      };
       
       handlers_.add(popup_.addAttachHandler(new AttachEvent.Handler()
       {
@@ -109,9 +97,27 @@ public class MathJax
          public void onAttachOrDetach(AttachEvent event)
          {
             if (!event.isAttached())
+            {
                detachHandlers();
+               return;
+            }
+            
+            MathJaxLoader.withMathJaxLoaded(new MathJaxLoader.Callback()
+            {
+               @Override
+               public void onLoaded(boolean alreadyLoaded)
+               {
+                  renderLatex();
+               }
+            });
          }
       }));
+   }
+   
+   public void renderLatex()
+   {
+      final List<Range> ranges = MathJaxUtil.findLatexChunks(docDisplay_);
+      renderQueue_.enqueueAndRender(ranges);
    }
    
    public void renderLatex(Range range)
@@ -120,6 +126,43 @@ public class MathJax
    }
    
    public void renderLatex(Range range, boolean background)
+   {
+      renderLatex(range, background, null);
+   }
+   
+   public void renderLatex(final Range range,
+                           final boolean background,
+                           final MathJaxTypesetCallback callback)
+   {
+      MathJaxLoader.withMathJaxLoaded(new MathJaxLoader.Callback()
+      {
+         @Override
+         public void onLoaded(boolean alreadyLoaded)
+         {
+            renderLatexImpl(range, background, callback);
+         }
+      });
+   }
+   
+   public void promotePopupToLineWidget()
+   {
+      if (range_ == null)
+         return;
+      
+      renderLatex(range_, false);
+   }
+   
+   // Private Methods ----
+   
+   @Inject
+   private void initialize(UIPrefs uiPrefs)
+   {
+      uiPrefs_ = uiPrefs;
+   }
+   
+   private void renderLatexImpl(final Range range,
+                                final boolean background,
+                                final MathJaxTypesetCallback callback)
    {
       String text = docDisplay_.getTextForRange(range);
       String pref = uiPrefs_.showLatexPreviewOnCursorIdle().getGlobalValue();
@@ -152,7 +195,7 @@ public class MathJax
             }
          }
 
-         renderLatexLineWidget(range, text);
+         renderLatexLineWidget(range, text, callback);
          return;
       }
       
@@ -160,7 +203,7 @@ public class MathJax
       // (don't reset render state)
       if (popup_.isShowing())
       {
-         renderPopup(text);
+         renderPopup(text, callback);
          return;
       }
       
@@ -168,27 +211,12 @@ public class MathJax
       range_ = range;
       anchor_ = docDisplay_.createAnchoredSelection(range.getStart(), range.getEnd());
       lastRenderedText_ = "";
-      popup_.setRenderToolbarVisible(isLatexChunk);
-      renderPopup(text);
+      renderPopup(text, callback);
    }
    
-   public void promotePopupToLineWidget()
-   {
-      if (range_ == null)
-         return;
-      
-      renderLatex(range_, false);
-   }
-   
-   // Private Methods ----
-   
-   @Inject
-   private void initialize(UIPrefs uiPrefs)
-   {
-      uiPrefs_ = uiPrefs;
-   }
-   
-   private void renderLatexLineWidget(final Range range, final String text)
+   private void renderLatexLineWidget(final Range range,
+                                      final String text,
+                                      final MathJaxTypesetCallback callback)
    {
       // end a previous render session if necessary (e.g. if a popup is showing)
       endRender();
@@ -225,6 +253,10 @@ public class MathJax
                   Element ppElement = el.getParentElement().getParentElement();
                   ppElement.getStyle().setHeight(height, Unit.PX);
                   docDisplay_.onLineWidgetChanged(lineWidget);
+                  
+                  // invoke supplied callback
+                  if (callback != null)
+                     callback.onMathJaxTypesetComplete(error);
                }
             });
          }
@@ -302,7 +334,8 @@ public class MathJax
       lastRenderedText_ = "";
    }
    
-   private void renderPopup(final String text)
+   private void renderPopup(final String text,
+                            final MathJaxTypesetCallback callback)
    {
       // no need to re-render if text hasn't changed or is empty
       if (text.equals(lastRenderedText_))
@@ -319,7 +352,7 @@ public class MathJax
       // just typeset
       if (popup_.isShowing())
       {
-         mathjaxTypeset(popup_.getContentElement(), text);
+         mathjaxTypeset(popup_.getContentElement(), text, callback);
          return;
       }
       
@@ -328,7 +361,24 @@ public class MathJax
       popup_.show();
       
       // typeset and position after typesetting finished
-      mathjaxTypeset(popup_.getContentElement(), text, onPopupTypeset_);
+      mathjaxTypeset(popup_.getContentElement(), text, new MathJaxTypesetCallback()
+      {
+         
+         @Override
+         public void onMathJaxTypesetComplete(boolean error)
+         {
+            // re-position popup after render
+            int offsetWidth  = popup_.getOffsetWidth();
+            int offsetHeight = popup_.getOffsetHeight();
+            ScreenCoordinates coordinates = computePopupPosition(offsetWidth, offsetHeight);
+            popup_.setPopupPosition(coordinates.getPageX(), coordinates.getPageY());
+            popup_.show();
+            
+            // invoke user callback if provided
+            if (callback != null)
+               callback.onMathJaxTypesetComplete(error);
+         }
+      });
    }
    
    private void endRender()
@@ -414,16 +464,9 @@ public class MathJax
       Rectangle bounds = docDisplay_.getRangeBounds(range_);
       Point center = bounds.center();
       
-      int pageX = bounds.getRight() + 10;
-      int pageY = center.getY() - (offsetHeight / 2);
-      
-      // check for overflow -- in such a case, display below rather than beside
-      int windowWidth = Window.getClientWidth();
-      if (pageX + offsetWidth >= (windowWidth - 10))
-      {
-         pageX = center.getX() - (offsetWidth / 2);
-         pageY = bounds.getBottom() + 10;
-      }
+      // prefer displaying popup below associated text
+      int pageX = center.getX() - (offsetWidth / 2);
+      int pageY = bounds.getBottom() + 10;
       
       return ScreenCoordinates.create(pageX, pageY);
    }
@@ -443,7 +486,7 @@ public class MathJax
    private final DocDisplay docDisplay_;
    private final MathJaxPopupPanel popup_;
    private final MathJaxBackgroundRenderer bgRenderer_;
-   private final MathJaxTypesetCallback onPopupTypeset_;
+   private final MathJaxRenderQueue renderQueue_;
    private final List<HandlerRegistration> handlers_;
    private final SafeMap<ChunkOutputWidget, PinnedLineWidget> cowToPlwMap_;