+++ src/gwt/src/org/rstudio/studio/client/common/mathjax/MathJaxBackgroundRenderer.java
@@ -19,18 +19,21 @@ import java.util.List;
 
 import org.rstudio.studio.client.common.mathjax.display.MathJaxPopupPanel;
 import org.rstudio.studio.client.workbench.views.source.editors.text.DocDisplay;
+import org.rstudio.studio.client.workbench.views.source.editors.text.ace.Position;
 import org.rstudio.studio.client.workbench.views.source.editors.text.ace.Range;
 import org.rstudio.studio.client.workbench.views.source.editors.text.events.CursorChangedEvent;
 import org.rstudio.studio.client.workbench.views.source.editors.text.events.CursorChangedHandler;
-import org.rstudio.studio.client.workbench.views.source.editors.text.events.EditorModeChangedEvent;
-
+import com.google.gwt.dom.client.NativeEvent;
 import com.google.gwt.event.dom.client.BlurEvent;
 import com.google.gwt.event.dom.client.BlurHandler;
 import com.google.gwt.event.dom.client.FocusEvent;
 import com.google.gwt.event.dom.client.FocusHandler;
 import com.google.gwt.event.logical.shared.AttachEvent;
 import com.google.gwt.event.shared.HandlerRegistration;
+import com.google.gwt.user.client.Event;
 import com.google.gwt.user.client.Timer;
+import com.google.gwt.user.client.Event.NativePreviewEvent;
+import com.google.gwt.user.client.Event.NativePreviewHandler;
 
 public class MathJaxBackgroundRenderer
 {
@@ -56,19 +59,24 @@ public class MathJaxBackgroundRenderer
          }
       };
       
+      mouseIdleTimer_ = new Timer()
+      {
+         @Override
+         public void run()
+         {
+            Position position = docDisplay_.screenCoordinatesToDocumentPosition(pageX_, pageY_);
+            Range range = MathJaxUtil.getLatexRange(docDisplay_, position);
+            if (range == null)
+               return;
+            
+            mathjax_.renderLatex(range, true);
+         }
+      };
+      
       handlers_.add(docDisplay_.addFocusHandler(new FocusHandler()
       {
          @Override
          public void onFocus(FocusEvent event)
-         {
-            beginMonitoring();
-         }
-      }));
-      
-      handlers_.add(docDisplay_.addEditorModeChangedHandler(new EditorModeChangedEvent.Handler()
-      {
-         @Override
-         public void onEditorModeChanged(EditorModeChangedEvent event)
          {
             beginMonitoring();
          }
@@ -89,17 +97,36 @@ public class MathJaxBackgroundRenderer
          public void onAttachOrDetach(AttachEvent event)
          {
             if (!event.isAttached())
-               resetHandlers();
+               onDetach();
          }
       }));
+      
    }
    
    private void beginMonitoring()
    {
-      resetHandlers();
+      endMonitoring();
       String id = docDisplay_.getModeId();
       if (!id.equals("mode/rmarkdown"))
          return;
+      
+      previewHandler_ = Event.addNativePreviewHandler(new NativePreviewHandler()
+      {
+         @Override
+         public void onPreviewNativeEvent(NativePreviewEvent preview)
+         {
+            if (preview.getTypeInt() != Event.ONMOUSEMOVE)
+            {
+               mouseIdleTimer_.cancel();
+               return;
+            }
+            
+            NativeEvent event = preview.getNativeEvent();
+            pageX_ = event.getClientX();
+            pageY_ = event.getClientY();
+            mouseIdleTimer_.schedule(700);
+         }
+      });
       
       cursorChangedHandler_ = docDisplay_.addCursorChangedHandler(new CursorChangedHandler()
       {
@@ -119,14 +146,11 @@ public class MathJaxBackgroundRenderer
    
    private void endMonitoring()
    {
-      resetHandlers();
-   }
-   
-   private void resetHandlers()
-   {
-      for (HandlerRegistration handler : handlers_)
-         handler.removeHandler();
-      handlers_.clear();
+      if (previewHandler_ != null)
+      {
+         previewHandler_.removeHandler();
+         previewHandler_ = null;
+      }
       
       if (cursorChangedHandler_ != null)
       {
@@ -135,11 +159,24 @@ public class MathJaxBackgroundRenderer
       }
    }
    
+   private void onDetach()
+   {
+      endMonitoring();
+      for (HandlerRegistration handler : handlers_)
+         handler.removeHandler();
+      handlers_.clear();
+   }
+   
    private final MathJax mathjax_;
    private final MathJaxPopupPanel popup_;
    private final DocDisplay docDisplay_;
    private final List<HandlerRegistration> handlers_;
    private final Timer renderTimer_;
    
+   private HandlerRegistration previewHandler_;
    private HandlerRegistration cursorChangedHandler_;
+   
+   private int pageX_;
+   private int pageY_;
+   private final Timer mouseIdleTimer_;
 }