+++ src/gwt/src/org/rstudio/studio/client/workbench/views/source/editors/text/rmd/ChunkContextToolbar.java
@@ -73,7 +73,6 @@ public class ChunkContextToolbar extends Composite
 
       initRunPrevious(dark);
       setRunPrevious(runPrevious);
-      initChangeChunkEngine(engine);
       
       initRun();
       setRun(run);
@@ -204,6 +203,7 @@ public class ChunkContextToolbar extends Composite
       return menuItem;
    }
 
+   @SuppressWarnings("unused")
    private void initChangeChunkEngine(String engine)
    {
       chunkTypeLabel_.setText(engine);